src/requirements.txt
asn1crypto==1.4.0
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
flask-talisman==0.8.1
Flask-WTF==0.15.1
greenlet==1.1.0
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
Pillow==8.3.1
pip-licenses==3.5.1
PTable==0.9.2
pyaxmlparser==0.3.26
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.20
typing-extensions==3.10.0.0
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.5.0
